Curated US politics RSS feeds

A verified collection of non-government RSS and Atom feeds covering US politics. The collection includes general political news, Congress, elections, polling, public opinion, and state politics. Feed activity was checked on the date recorded in verified_at; publisher availability can change after that date. Access it directly as textpress::rss_politics.

rss_politics

Format

A data frame with one row per feed and seven variables:

source

Publisher or organization name.

feed

Publisher-supplied or descriptive feed name.

category

Primary subject category assigned by textpress.

source_type

Type of publisher or organization.

url

RSS or Atom feed URL.

verified_at

UTC date and time when the feed was last checked.

latest_item_at

UTC publication date and time of the newest item observed during verification.

Source

Publisher-provided RSS and Atom endpoints.

Examples

head(textpress::rss_politics)
#>             source     feed  category    source_type
#> 1         POLITICO Congress  congress           news
#> 2   Punchbowl News      All  congress           news
#> 3        Roll Call All news  congress           news
#> 4         The Hill    House  congress           news
#> 5         The Hill   Senate  congress           news
#> 6 Ballotpedia News      All elections nonprofit_news
#>                                         url             verified_at
#> 1     https://rss.politico.com/congress.xml 2026-08-26 03:06:19 UTC
#> 2              https://punchbowl.news/feed/ 2026-08-26 03:06:19 UTC
#> 3                https://rollcall.com/feed/ 2026-08-26 03:06:19 UTC
#> 4  https://thehill.com/homenews/house/feed/ 2026-08-26 03:06:19 UTC
#> 5 https://thehill.com/homenews/senate/feed/ 2026-08-26 03:06:19 UTC
#> 6        https://news.ballotpedia.org/feed/ 2026-08-26 03:06:19 UTC
#>            latest_item_at
#> 1 2026-08-25 11:57:34 UTC
#> 2 2026-08-23 20:30:00 UTC
#> 3 2026-08-25 16:58:47 UTC
#> 4 2026-08-26 00:42:43 UTC
#> 5 2026-08-26 03:14:22 UTC
#> 6 2026-08-25 20:00:00 UTC